% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendMetabolomicsWorkbench.R
\name{MsBackendMetabolomicsWorkbench}
\alias{MsBackendMetabolomicsWorkbench}
\alias{MsBackendMetabolomicsWorkbench-class}
\alias{backendInitialize,MsBackendMetabolomicsWorkbench-method}
\alias{backendRequiredSpectraVariables,MsBackendMetabolomicsWorkbench-method}
\alias{mwb_sync}
\title{MsBackend representing MS data from Metabolomics Workbench}
\usage{
MsBackendMetabolomicsWorkbench()

\S4method{backendInitialize}{MsBackendMetabolomicsWorkbench}(
  object,
  mwbId = character(),
  filePattern = "mzML$|CDF$|cdf$|mzXML$",
  fileName = character(),
  ftp_zip = FALSE,
  offline = FALSE,
  ...
)

\S4method{backendRequiredSpectraVariables}{MsBackendMetabolomicsWorkbench}(object, ...)

mwb_sync(x, offline = FALSE)
}
\arguments{
\item{object}{an instance of \code{MsBackendMetabolomicsWorkbench}.}

\item{mwbId}{\code{character(1)} with the ID of a single
MetabolomicsWorkbench data set/experiment.}

\item{filePattern}{\code{character} with the pattern defining the
supported (or requested) file types. Defaults to
\code{filePattern = "mzML$|CDF$|cdf$|mzXML$"} hence restricting to
mzML, CDF and mzXML files which are supported by \emph{Spectra}'s
\code{MsBackendMzR} backend.}

\item{fileName}{\code{character} defining the names of specific data files of a
data set that should be downloaded and cached.}

\item{ftp_zip}{for \code{mwb_sync_data_files()}: \code{logical(1)} download the
complete zip of the experiment from the FTP server. Defaults to \code{FALSE},
in which case the files are downloaded singularly via POST request.}

\item{offline}{\code{logical(1)} whether only locally cached content
should be evaluated/loaded.}

\item{...}{additional parameters; currently ignored.}

\item{x}{an instance of \code{MsBackendMetabolomicsWorkbench}.}
}
\value{
\itemize{
\item For \code{MsBackendMetabolomicsWorkbench()}: an instance of
\code{MsBackendMetabolomicsWorkbench}.
\item For \code{backendInitialize()}: an instance of
\code{MsBackendMetabolomicsWorkbench} with the MS data of the specified
MetabolomicsWorkbench data set.
\item For \code{backendRequiredSpectraVariables()}: \code{character} with spectra
variables that are needed for the backend to provide the MS data.
\item For \code{mwb_sync()}: the input \code{MsBackendMetabolomicsWorkbench} with
the paths to the locally cached data files being eventually
updated.
}
}
\description{
\code{MsBackendMetabolomicsWorkbench} retrieves and represents mass spectrometry
(MS) data from metabolomics studies stored in the
\href{https://metabolomicsworkbench.org/}{Metabolomics Workbench} repository, a
data resource developed by the NIH Common Fund's Data Repository and
Coordinating Center (DRCC) at the San Diego Supercomputer Center, University
of California San Diego.
The repository provides access to study metadata, processed experimental
results, metabolite structures, and reference compound information through a
RESTful HTTP API / FTP server / POST request. The backend directly extends
the \link[Spectra:MsBackendMzR]{Spectra::MsBackendMzR} backend from the \emph{Spectra} package and hence
supports MS data in mzML, CDF, and mzXML format. Data in other formats cannot
be loaded with \code{MsBackendMetabolomicsWorkbench}. Upon initialization with the
\code{backendInitialize()} method, the \code{MsBackendMetabolomicsWorkbench} backend
fetches and caches study data files locally using Bioconductor's
\emph{BiocFileCache} package, avoiding repeated queries to the remote repository.
See the help and vignettes of that package for details on cached data
resources. Additional utility functions for management of cached files are
also provided by \emph{MsBackendMetabolomicsWorkbench}. See help for
\code{\link[=mwb_cached_data_files]{mwb_cached_data_files()}} for more information.
}
\details{
The backend uses the
\href{https://bioconductor.org/packages/BiocFileCache}{BiocFileCache} package for
caching of the data files. These are stored in the default local
\emph{BiocFileCache} cache along with additional metadata that includes the
Metabolomics Workbench ID. Note that at present only MS data files in \emph{mzML},
\emph{CDF} and \emph{mzXML} format are supported.

The \code{MsBackendMetabolomicsWorkbench} backend defines and provides additional
spectra variables \code{"mwb_id"}, \code{"zip_file"} and \code{"file_name"} that list
the MetabolomicsWorkbench ID, the original zip file name and the original
data file name on the Metabolomics Workbench ftp server for each individual
spectrum. The \code{"file_name"} can be used for the mapping between the
experiment's samples and the individual data files, respective their spectra.

The \code{MsBackendMetabolomicsWorkbench} backend is considered \emph{read-only} and
does thus not support changing \emph{m/z} and intensity values directly.
}
\note{
To account for transient network failures and high server load on the
Metabolomics Workbench endpoint, download functions automatically retry
failed requests. An error is raised after 5 consecutive failed attempts.
Between each attempt, the function waits for a progressively increasing time
period (5 seconds between the first and second attempt, 10 seconds between
the second and third, and so forth). The sleep time multiplier can be
configured via the \code{"mwb.sleep_mult"} option (defaults to \code{5}). An active
internet connection is required for all non-cached operations; use
\code{offline = TRUE} in \code{backendInitialize()} to suppress remote requests and
rely exclusively on the local \emph{BiocFileCache} cache.
}
\section{Initialization and loading of data}{


New instances of the class can be created with the
\code{MsBackendMetabolomicsWorkbench()} function. Data is loaded and initialized
using the \code{backendInitialize()} function, which accepts parameters \code{mwbId},
\code{filePattern} and \code{ftp_zip}. \code{mwbId} must be the accession of a \strong{single}
existing Metabolomics Workbench study (e.g. \code{"ST000016"}). Optional parameter
\code{filePattern} defines the pattern used to filter the file names of the MS
data files and defaults to data files with file endings of supported MS data
formats. Optional parameter \code{ftp_zip = TRUE} will download the complete zip
file of the experiment from the FTP server and extract the data files
locally, which can be faster than downloading the files individually via POST
request. However if only a subset of the data files is required, it is more
efficient to download the files separately \emph{via} POST request with
\code{ftp_zip = FALSE} and \code{filePattern} set to the desired file name pattern.
\code{backendInitialize()} requires an active internet connection, as the function
queries the Metabolomics Workbench via POST request and compares remote file
content against locally cached files before synchronizing any changes or
updates. This behavior can be bypassed with \code{offline = TRUE}, in which case
only locally cached content is used.

The \code{backendRequiredSpectraVariables()} function returns the names of the
spectra variables required for the backend to provide the MS data.

The \code{mwb_sync()} function can be used to \emph{synchronize} the local data cache
and ensure that all study data files are locally available. The function
checks the local cache and downloads any missing data files from the
Metabolomics Workbench repository.
}

\examples{

library(MsBackendMetabolomicsWorkbench)

## List files of a MetabolomicsWorkbench data set
mwb_list_files("ST002115")

## Initialize a MsBackendMetabolomicsWorkbench representing all MS
## data files of the data set with the ID "ST002115". This will
## download and cache all files and subsequently load and represent
## them in R.

be <- backendInitialize(MsBackendMetabolomicsWorkbench(),
                        "ST002115",
                        filePattern = "DMSO_01_RP.mzXML$")
be

## The `mwb_sync()` function can be used to ensure that all data
## files are available locally. This function will eventually download
## missing data files or update their paths.
be <- mwb_sync(be)
}
\author{
Gabriele Tomè, Philippine Louail, Johannes Rainer
}
